Base Config Entities.
Each extend ConfigEntity for the optional utility of the automatic name prop from file name (be careful, as multiple derived classes will get the name of the first dervived class - so name will need to be passed via super), also use of a mixin in the inheritance chain will add an extra level so forcing the same behaviour as multiple level inheritance.

Watchers contain Items (one for each instance of an item, i.e. gpu item will have 2 item objects if 2 instances are configured to be watched),
Items contain the measures to track on that item. i.e. temporate
A measure on an item then contains Records for each measure the source knows about. i.e core 1 temp, core 2 temp etc.
A Record is just a key value pair object of the value and the values identifier (name).