Aggregate and HistoryCount added with MeasureConfiguration class.
Aggregates are per a record. Once an aggregate has been added to a record it cannot be overriden.  The measure is responsible for setting the aggregates and calling the calculate function on each aggregate when setting the records on its self.
Calculate function takes a series of useful values and should set the _value prop on the aggregates class with the worked out aggregate value. Calc gets current records value, previous records value and the previous aggregate object.
aggregates also have a dataCount prop which is used to keep track of the number of record values that have been calculated in the aggregate, the base aggregate's calculate method sets this value, which should be called from derived calculate methods if required.

Min aggregate class is a solid aggregate class which works out the minimum value for a certain record.

A limit on historic data of records is set via the historyCount value, which will remove older records above the limit. Note: aggregates are calculated before the previous record is removed (if required) so even with a history count of 1 will still mean the aggregate can work on the current and previous records.

MeasureConfiguration. measures are still instantiated with arguments but a configuration object is now present to bundle together the various configuration values throught the object hierarchy. So watchers and items now take a MeasureConfiguration which contains the measure names for the item, the history count value for the measures and the aggregates for the measures of the items.

Updated config.json with new nodes and configuration/watcherfactories/watchers with new historyCount/MeasureConfiguration class api.

populated method on Item now correctly checks the measures records (the latest record batch) for populated state.

To make the above work, items are not recreated for each watcher run, so when items are set in Watcher class they are deduped so only newly found item instances are added to the known items for the watcher. Its also possible to remove all Items via the clearItems method, for resetting purposes.